feat: 개설학과 정보 관련 Database 로직 추가
- 개설학과 정보 저장을 위한 OpenMajorDatabase 및 관련 클래스(Dao, Entity, Factory, Constructor)를 추가했습니다.
- 기존 Timetable 관련 Database 로직을 timetable 패키지로 이동했습니다.
- DatabaseModule에 OpenMajorDatabase 관련 의존성을 추가했습니다.
