--- composeApp/src/commonMain/kotlin/com/chukchukhaksa/mobile/local/common/database/database/DatabaseName.kt
@@ -1,7 +1,0 @@
-package com.chukchukhaksa.mobile.local.common.database.database
-
-object DatabaseName {
-    const val OPEN_MAJOR = "open-major-database"
-    const val TIMETABLE = "timetable-list-database"
-    const val OPEN_LECTURE = "open-lecture-database"
-}
