--- composeApp/src/commonMain/kotlin/com/chukchukhaksa/mobile/local/common/database/entity/TimetableEntity.kt
@@ -1,14 +1,0 @@
-package com.chukchukhaksa.mobile.local.common.database.entity
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import com.chukchukhaksa.mobile.common.model.TimetableCell
-
-@Entity
-data class TimetableEntity(
-    @PrimaryKey val createTime: Long,
-    var year: String,
-    var semester: String,
-    var name: String,
-    var cellList: List<TimetableCell>,
-)
