+++ composeApp/src/commonMain/kotlin/com/chukchukhaksa/mobile/local/database/common/DatabaseName.kt
@@ -1,0 +1,7 @@
+package com.chukchukhaksa.mobile.local.database.common
+
+object DatabaseName {
+    const val OPEN_MAJOR = "open-major-database"
+    const val TIMETABLE = "timetable-list-database"
+    const val OPEN_LECTURE = "open-lecture-database"
+}
