--- composeApp/src/iosMain/kotlin/com/chukchukhaksa/mobile/local/common/database/database/TimetableDatabaseFactory.ios.kt
@@ -1,32 +1,0 @@
-package com.chukchukhaksa.mobile.local.common.database.database
-
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import androidx.sqlite.driver.bundled.BundledSQLiteDriver
-import kotlinx.cinterop.ExperimentalForeignApi
-import platform.Foundation.NSDocumentDirectory
-import platform.Foundation.NSFileManager
-import platform.Foundation.NSUserDomainMask
-
-actual class TimetableDatabaseFactory {
-    actual fun create(): RoomDatabase.Builder<TimetableDatabase> {
-        val dbFilePath = documentDirectory() + "/${DatabaseName.TIMETABLE}"
-        return Room
-            .databaseBuilder<TimetableDatabase>(
-                name = dbFilePath
-            ).setDriver(BundledSQLiteDriver())
-    }
-
-    @OptIn(ExperimentalForeignApi::class)
-    private fun documentDirectory(): String {
-        val documentDirectory =
-            NSFileManager.defaultManager.URLForDirectory(
-                directory = NSDocumentDirectory,
-                inDomain = NSUserDomainMask,
-                appropriateForURL = null,
-                create = false,
-                error = null,
-            )
-        return requireNotNull(documentDirectory?.path)
-    }
-}
